choruses
added sine chorus
change mix to float for both choruses
added delay line handling functions

+++ Inc/audio/simpleChorus.h
@@ -27,7 +27,7 @@ typedef struct
     float delayBuffer[SIMPLE_CHORUS_DELAY_SIZE];
     int16_t frequency; // in Hz/100, max 9375 --> 93.75 Hz
     int16_t depth; // 0 to 255
-    int16_t mix; // 0 to 255
+    float mix; // 0.0 to 1.0
     int16_t lfoVal; // from -255 to +255
     int16_t lfoValOld;
     uint8_t lfoQuadrant; // 0 (ascending) or 1 (descending)
